loading offset is long

+++ Lazinator/Buffers/IMemoryChunkReference.cs
@@ -6,7 +6,7 @@ namespace Lazinator.Buffers
     public interface IMemoryChunkReference
     {
         int MemoryChunkID { get; set; }
-        int OffsetForLoading { get; set; }
+        long OffsetForLoading { get; set; }
         int PreTruncationLength { get; set; }
         int AdditionalOffset { get; set; }
         int FinalLength { get; set; }
